Refactored main view to be more modular and more memoization
