+++ atlas-extension/project-atlas/package.json
@@ -48,6 +48,7 @@
         "@types/node": "^12.20.12",
         "@types/react": "^17.0.5",
         "@types/react-dom": "^17.0.3",
+        "@types/styled-components": "^5.1.9",
         "@types/vscode": "^1.55.0",
         "@typescript-eslint/eslint-plugin": "^4.9.0",
         "@typescript-eslint/parser": "^4.9.0",
@@ -67,7 +68,6 @@
     },
     "dependencies": {
         "@svgdotjs/svg.js": "^3.0.16",
-        "@types/styled-components": "^5.1.9",
         "react": "^17.0.1",
         "react-dom": "^17.0.1",
         "react-icons": "^4.2.0",
