+++ atlas-extension/project-atlas/src/app/AtlasApp.tsx
@@ -1,16 +1,19 @@
 import { MainView } from "./components/MainView";
 import CameraContainer from "./containers/CameraContainer";
 import MessagingContainer from "./containers/MessagingContainer";
+import ModalsContainer from "./containers/ModalsContainer";
 import ProjectContainer from "./containers/ProjectContainer";
 
 export function AtlasApp() {
     return (
-        <ProjectContainer>
-            <CameraContainer>
-                <MessagingContainer>
-                    <MainView />
-                </MessagingContainer>
-            </CameraContainer>
-        </ProjectContainer>
+        <ModalsContainer>
+            <ProjectContainer>
+                <CameraContainer>
+                    <MessagingContainer>
+                        <MainView />
+                    </MessagingContainer>
+                </CameraContainer>
+            </ProjectContainer>
+        </ModalsContainer>
     );
 }
