+++ atlas-extension/project-atlas/src/app/models/messages.ts
@@ -5,12 +5,10 @@ export enum WebViewMessageType {
     Refresh = "REFRESH",
     ChangeSource = "CHANGE_SOURCE",
     UnloadProject = "UNLOAD_PROJECT",
-    None = "NONE",
 }
 
 export enum APIMessageType {
     NewJSONData = "NEW_JSON_DATA",
-    None = "NONE",
 }
 
 interface IMessage {
@@ -32,13 +30,6 @@ export interface ChangeSourceMessage extends IMessage {
 export interface UnloadProjectMessage extends IMessage {
     type: WebViewMessageType.UnloadProject;
 }
-
-export interface NoMessage extends IMessage {
-    type: WebViewMessageType.None | APIMessageType.None;
-}
-export interface NewMessage extends IMessage {
-    type: WebViewMessageType.None;
-}
 export interface NewJSONDataMessage extends IMessage {
     type: APIMessageType.NewJSONData;
     data: ProjectJSON | null;
@@ -48,9 +39,8 @@ export type WebViewMessage =
     | DocumentLoadedMessage
     | RefreshMessage
     | ChangeSourceMessage
-    | UnloadProjectMessage
-    | NoMessage; // TODO: TEMP
+    | UnloadProjectMessage;
 
-export type APIMessage = NewJSONDataMessage | NoMessage;
+export type APIMessage = NewJSONDataMessage;
 
 export type MessageType = WebViewMessage | APIMessage;
